refactor: 챌린지 관련 타입들 분리
