+++ types/lecture.ts
@@ -35,21 +35,7 @@ export interface LectureData {
   upload_type: number;
 }
 
-export interface ChallengeResponse {
-  challenge_id: number;
-  Challenges: {
-    id: number;
-    name: string;
-  };
-}
-
-export interface Challenge {
-  id: number;
-  name: string;
-  lecture_num?: number;
-}
-
-export interface ChallengeLectureResponse {
+export interface ChallengeLectures {
   lecture_id: number;
   sequence: number;
   Lectures: {
@@ -64,15 +50,6 @@ export interface ChallengeLectureResponse {
       title: string;
       contents: string;
     }>;
-  };
-}
-
-export interface ChallengeUser {
-  challenge_id: number;
-  Challenges: {
-    id: number;
-    open_date: string;
-    close_date: string;
   };
 }
 
@@ -110,11 +87,6 @@ export interface LectureWithSequence {
   sequence: number;
   assignment_title: string;
   assignment: string;
-}
-
-export interface ChallengeOrder {
-  challengeId: number;
-  order: number;
 }
 
 export interface LectureFormInitialData {
